+++ .env
@@ -25,9 +25,6 @@ export CPEM_VERSION=v3.6.0
 export KUBERNETES_VERSION=v1.26.2
 export TALOS_VERSION=v1.3.6
 
-## Colima (MacOs)
-export DOCKER_HOST=unix://${HOME}/.colima/default/docker.sock
-
 ## Project specific
 source .venv/bin/activate
 source <(talosctl completion zsh)
@@ -36,3 +33,4 @@ export TALOSCONFIG=secrets/talosconfig
 export TOEM_SECRETS_DIR="secrets"
 export TOEM_MANIFEST_DIR="${TOEM_SECRETS_DIR}/manifest"
 export TOEM_PROJECT_ROOT="$(git rev-parse --show-toplevel)"
+export TOEM_TEST_SUBDOMAIN="v2"
