+++ apps/dns-and-tls-dependencies/Chart.lock
@@ -1,0 +1,6 @@
+dependencies:
+- name: cert-manager
+  repository: https://charts.jetstack.io
+  version: v1.11.0
+digest: sha256:87ce75f397241680f6d6840d0200a5e5f7030ce41429fa87b1fd883829339a1f
+generated: "2023-04-06T17:36:57.262779+02:00"
