+++ apps/dns-and-tls-dependencies/Chart.yaml
@@ -1,0 +1,12 @@
+apiVersion: v2
+name: dns-and-tls-dependencies
+description: Chart tracks compatible versions of talos + cilium + external-dns + cert-manager
+
+type: application
+version: 0.1.0
+appVersion: "1.16.0"
+
+dependencies:
+  - name: cert-manager
+    version: v1.11.0
+    repository: https://charts.jetstack.io
