+++ apps/dns-and-tls-dependencies/templates/rbac.yaml
@@ -1,0 +1,52 @@
+---
+# Source: external-dns/templates/serviceaccount.yaml
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: external-dns-gcp
+  namespace: {{ .Release.Namespace }}
+  labels:
+    app.kubernetes.io/name: external-dns-gcp
+    app.kubernetes.io/instance: external-dns-gcp
+    app.kubernetes.io/version: {{ .Values.external_dns.version }}
+---
+# Source: external-dns/templates/clusterrole.yaml
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: external-dns-gcp
+  labels:
+    app.kubernetes.io/name: external-dns-gcp
+    app.kubernetes.io/instance: external-dns-gcp
+    app.kubernetes.io/version: {{ .Values.external_dns.version }}
+rules:
+  - apiGroups: [""]
+    resources: ["nodes"]
+    verbs: ["list","watch"]
+  - apiGroups: [""]
+    resources: ["pods"]
+    verbs: ["get","watch","list"]
+  - apiGroups: [""]
+    resources: ["services","endpoints"]
+    verbs: ["get","watch","list"]
+  - apiGroups: ["extensions","networking.k8s.io"]
+    resources: ["ingresses"]
+    verbs: ["get","watch","list"]
+---
+# Source: external-dns/templates/clusterrolebinding.yaml
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: external-dns-gcp-viewer
+  labels:
+    app.kubernetes.io/name: external-dns-gcp
+    app.kubernetes.io/instance: external-dns-gcp
+    app.kubernetes.io/version: {{ .Values.external_dns.version }}
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: external-dns-gcp
+subjects:
+  - kind: ServiceAccount
+    name: external-dns-gcp
+    namespace: {{ .Release.Namespace }}
