+++ apps/dns-and-tls-dependencies/templates/service.yaml
@@ -1,0 +1,19 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: external-dns-gcp
+  namespace: {{ .Release.Namespace }}
+  labels:
+    app.kubernetes.io/name: external-dns-gcp
+    app.kubernetes.io/instance: external-dns-gcp
+    app.kubernetes.io/version: {{ .Values.external_dns.version }}
+spec:
+  type: ClusterIP
+  selector:
+    app.kubernetes.io/name: external-dns-gcp
+    app.kubernetes.io/instance: external-dns-gcp
+  ports:
+    - name: http
+      port: 7979
+      targetPort: http
+      protocol: TCP
