+++ apps/dns-and-tls/Chart.yaml
@@ -1,0 +1,7 @@
+apiVersion: v2
+name: dns-and-tls
+description: Chart tracks compatible versions of talos + cilium + external-dns + cert-manager
+
+type: application
+version: 0.1.0
+appVersion: "1.16.0"
