+++ apps/dns-and-tls/templates/ClusterIssuer.yaml
@@ -1,0 +1,21 @@
+{{- if .Values.letsencrypt.enabled }}
+apiVersion: cert-manager.io/v1
+kind: ClusterIssuer
+metadata:
+  name: letsencrypt-gcp
+spec:
+  acme:
+    email: {{ .Values.letsencrypt.email }}
+    server: https://acme-v02.api.letsencrypt.org/directory
+    privateKeySecretRef:
+      name: letsencrypt-issuer-account-key
+    solvers:
+      {{- if .Values.letsencrypt.google.enabled }}
+      - dns01:
+          cloudDNS:
+            project: {{ .Values.letsencrypt.google.project_id }}
+            serviceAccountSecretRef:
+              name: external-dns
+              key: credentials.json
+      {{ end }}
+{{ end }}
