+++ apps/network-services-dependencies/.gitignore
