+++ apps/network-services-dependencies/Chart.yaml
