+++ apps/network-services-dependencies/namespace.yaml
