+++ apps/network-services-dependencies/values.yaml
@@ -1,4 +1,9 @@
 cilium:
+  ingressController:
+    enabled: true
+  loadBalancer:
+    l7:
+      backend: envoy
   MTU: 9000
   bpf:
     masquerade: true
@@ -33,7 +38,7 @@ cilium:
     enabled: true
   ipv6:
     enabled: false # ToDo: true | https://yolops.net/k8s-dualstack-cilium.html
-  kubeProxyReplacement: disabled # ToDo: strick | Run without KubeProxy https://docs.cilium.io/en/stable/network/kubernetes/kubeproxy-free/#kube-proxy-hybrid-modes
+  kubeProxyReplacement: partial # ToDo: strick | Run without KubeProxy https://docs.cilium.io/en/stable/network/kubernetes/kubeproxy-free/#kube-proxy-hybrid-modes
   policyAuditMode: true
   policyEnforcementMode: default # ToDo: always | https://docs.cilium.io/en/stable/security/policy/intro/#policy-enforcement-modes
   cgroup:
