+++ apps/network-services/.gitignore
