+++ apps/network-services/templates/bgp.yaml
